fix(jest): Resolve module resolution error in tests
The `yarn test` command was failing with a `SyntaxError: The requested module 'chessgroundx' does not provide an export named 'Chessground'`.

This error was caused by a side effect of having a `sanitizeURL` function in `client/main.ts`, which uses browser-specific APIs like `URL` and `window.location.origin`. This seemed to interfere with Jest's module resolution mechanism, specifically for the `chessgroundx` package.

The fix involves moving the `sanitizeURL` function to its own file, `client/url.ts`, and updating the imports accordingly. This isolates the function and prevents it from affecting the test environment in unexpected ways.

Additionally, the Jest configuration in `package.json` was improved to:
- Add a `moduleNameMapper` for the `@/` alias.
- Add `transformIgnorePatterns` to correctly handle ES modules in `node_modules`.
- Fix an incorrect import of `Chessground` to point to the correct file within the `chessgroundx` package.
