+++ tsconfig.json
@@ -19,6 +19,7 @@
         "module": "es2020",
         "moduleResolution": "node",
         "lib": ["DOM", "es2020"],
-        "target": "es2020"
+        "target": "es2020",
+        "isolatedModules": true
     }
 }
